refactor: move refresh token grant to new file (#1187)
To make the `token.go` file less busy.
